other changes

+++ gfx/BufferedDisplay.cpp
@@ -7,7 +7,7 @@
 
   It may happen that there is not enough memory on the system to allocate BufferedDisplay.
   And, for example, on ESP32 this will result in a stack trace print to Serial and a system restart.
-//
+
 */
 
 #include "BufferedDisplay.h"
@@ -57,7 +57,7 @@ static const rgb_t default_4bit_palette[] = {
    * @y2     - height = y2 - y1
    *      this causes the memory (internal buffer) allocation width * height * MDT_SIZE
    */
-  BufferedDisplay::BufferedDisplay(const int16_t x1, const int16_t y1, const int16_t x2, const int16_t y2, const rgb_t aBgColor)
+  BufferedDisplay::BufferedDisplay(const int32_t x1, const int32_t y1, const int32_t x2, const int32_t y2, const rgb_t aBgColor)
     : TSD_SCREEN(x2 - x1, y2 - y1)
   {
     clip_t clip = {x1, y1, x2, y2};
@@ -124,7 +124,7 @@ static const rgb_t default_4bit_palette[] = {
    * x      - x (left)
    * y      - y (top)
    */
-  void BufferedDisplay::setPos(const int16_t x, const int16_t y)
+  void BufferedDisplay::setPos(const int32_t x, const int32_t y)
   {
     _clip.x2 = _clip.x2 - _clip.x1 + x;
     _clip.x1 = x;
@@ -135,7 +135,7 @@ static const rgb_t default_4bit_palette[] = {
   /**
    * writeAddrWindow - from tft interface, not to use by user
    */
-  void BufferedDisplay::writeAddrWindow(const int16_t x, const int16_t y, const int16_t w, const int16_t h)
+  void BufferedDisplay::writeAddrWindow(const int32_t x, const int32_t y, const int32_t w, const int32_t h)
   {
     addr_x = x - _clip.x1;
     addr_y = y - _clip.y1;
@@ -234,7 +234,7 @@ static const rgb_t default_4bit_palette[] = {
    * rotateRight - simple square bit to bit rotation in internal buffer
    *  from (x,y) position, d pixels size
    */
-  void BufferedDisplay::rotateRight(const int16_t x, const int16_t y, int16_t d)
+  void BufferedDisplay::rotateRight(const int32_t x, const int32_t y, int32_t d)
   {
     if (x < 0 || y < 0) {
       return;
@@ -263,7 +263,7 @@ static const rgb_t default_4bit_palette[] = {
    *  rotateLeft - simple square bit to bit rotation in internal buffer
    *   from (x,y) position, d pixels size
    */
-  void BufferedDisplay::rotateLeft(const int16_t x, const int16_t y, int16_t d)
+  void BufferedDisplay::rotateLeft(const int32_t x, const int32_t y, int32_t d)
   {
     if (x < 0 || y < 0) {
       return;
@@ -292,7 +292,7 @@ static const rgb_t default_4bit_palette[] = {
    *  flipHorizontal - simple bit to bit flip horizontal in internal buffer
    *   from (x,y) position, (w, h) pixels size
    */
-  void BufferedDisplay::flipHorizontal(const int16_t x, const int16_t y, int16_t w, int16_t h)
+  void BufferedDisplay::flipHorizontal(const int32_t x, const int32_t y, int32_t w, int32_t h)
   {
     if (x < 0 || y < 0) {
       return;
@@ -317,7 +317,7 @@ static const rgb_t default_4bit_palette[] = {
    * flipVertical - simple bit to bit flip vertical in internal buffer
    *  from (x,y) position, (w, h) pixels size
    */
-  void BufferedDisplay::flipVertical(const int16_t x, const int16_t y, int16_t w, int16_t h)
+  void BufferedDisplay::flipVertical(const int32_t x, const int32_t y, int32_t w, int32_t h)
   {
     if (x < 0 || y < 0) {
       return;
@@ -341,7 +341,7 @@ static const rgb_t default_4bit_palette[] = {
   /**
    * drawMDTBuffer - from tft interface, not to use by user
    */
-  void BufferedDisplay::drawMDTBuffer(int16_t x, int16_t y, int16_t w, int16_t h, const uint8_t* buffer)
+  void BufferedDisplay::drawMDTBuffer(int32_t x, int32_t y, int32_t w, int32_t h, const uint8_t* buffer)
   {
     if (!buf) {
       return;
@@ -397,7 +397,7 @@ static const rgb_t default_4bit_palette[] = {
 //    }
   }
 
-  rgb_t BufferedDisplay::readPixel(clip_t& cli, int16_t x, int16_t y)
+  rgb_t BufferedDisplay::readPixel(clip_t& cli, int32_t x, int32_t y)
   {
     if (x >= _clip.x1 && x < _clip.x2 && y >= _clip.y1 && y < _clip.y2 ) {
       return getColor(x, y);
